+++ src/Channel/Tests/TestCase.php
@@ -77,6 +77,7 @@ abstract class TestCase extends Orchestra
 
     protected function setUpDatabase($app)
     {
+        $this->loadMigrationsFrom(__DIR__ . '/migrations');
         \Artisan::call('migrate', ['--force' => true]);
     }