+++ Changelog.md
@@ -15,6 +15,7 @@
 - Added `currency` field to the Channel model/table
 - Added the `is_hidden` field to the `Property` model
 - Added the `withImages` method to the product search class (eager loads media)
+- Added the `channelables` table for being many-to-many polymorphic relationships with channels and arbitrary models
 
 ## 3.x Series